# AES and  RSA (bn_mul) optimizations for Visual Studio 64bit