--- a/include/mbedtls/aesni.h
+++ b/include/mbedtls/aesni.h
@@ -35,6 +35,11 @@
 #define MBEDTLS_HAVE_X86_64
 #endif
 
+#if defined(_MSC_VER) && defined(_M_X64) &&  \
+    ! defined(MBEDTLS_HAVE_X86_64)
+#define MBEDTLS_HAVE_X86_64
+#endif
+
 #if defined(MBEDTLS_HAVE_X86_64)
 
 #ifdef __cplusplus