@mrsshr 2018-08-23
AWESOME! :D
I have a question. Is AES-NI not supported in 32 bit mode?