@hanno-arm 2018-08-23
Hi @orlx,

thanks alot for your contribution! :+1: We will look into it and come back to you afterwards.

Kind regards,
Hanno