@despair86 2018-09-04
@mrsshr the inline assembler code for GCC and Clang appears to be 32-bit clean. I'll come back shortly with test results.